CircleCI cleanup
Move to the advanced dir, line up with latest orb example
