+++ advanced/approve_circleci_job/README.md
@@ -1,7 +1,7 @@
 # Approving a CircleCI Job with Sym
 This section illustrates how to use the [Sym Orb](https://circleci.com/developer/orbs/orb/sym/sym) in your CircleCI pipeline and how to approve a CircleCI job with hooks
 
-A diff between this example and the basic [Approval](../approvals) example: [Diff](https://github.com/symopsio/examples/compare/c6453075e3a1d10a7a80b9ec55f0dc5a516044e4...1d9fcbf54d579331a54b8ad0ac9dc5548a30fda7)
+A diff between this example and the basic [Approval](../../approvals) example: [Diff](https://github.com/symopsio/examples/compare/c6453075e3a1d10a7a80b9ec55f0dc5a516044e4...1d9fcbf54d579331a54b8ad0ac9dc5548a30fda7)
 
 ## Tutorial
 
@@ -21,33 +21,35 @@ Once approved, Sym will automatically resume the workflow after the `wait_for_sy
 orbs:
   sym: sym/sym@1.0.0
 
-workflows:
-  main:
-    jobs:
-      # This will start the Sym flow
-      - sym/request:
-          flow_srn: sym:flow:ci-approval-prod:latest
-          flow_inputs: '{
-                      "workflow_url": "${CIRCLE_BUILD_URL}",
-                      "merging_user": "${CIRCLE_USERNAME}",
-                      "workflow_id": "${CIRCLE_WORKFLOW_ID}"
-                  }'
-          requires:
-            - terraform_acceptance_test
-          context: sym-bot-token
+jobs:
+  prod-deploy:
+    docker:
+      - image: cimg/base:stable
+    steps:
+      - run:
+          command: echo "Deploying to prod"
 
-      # The workflow will pause here and wait for approval.
-      # Once approved, Sym will approve this job to continue the workflow
+workflows:
+  sym-approval-example:
+    jobs:
+      - sym/request:
+          flow_srn: "healthy-health:flow:approval-ci-deploy-prod:latest"
+          flow_inputs: '{
+                          "build_url": "${CIRCLE_BUILD_URL}",
+                          "workflow_id": "${CIRCLE_WORKFLOW_ID}"
+                        }'
+          request_slug: "prod-deploy"
       - wait_for_sym_approval:
           type: approval
           requires:
             - sym/request
-
-      # After approving the Sym request in Slack, CircleCI will continue to this job
-      - deploy_prod:
+      - sym/validate:
+          request_slug: "prod-deploy"
           requires:
             - wait_for_sym_approval
-          ...
+      - prod-deploy:
+          requires:
+            - sym/validate
 ```
 
 For more information of the Sym CircleCI orb, please check out [our docs](https://circleci.com/developer/orbs/orb/sym/sym).
