+++ advanced/approve_circleci_job/versions.tf
@@ -1,5 +1,10 @@
 terraform {
+  required_version = ">= 1.0.0, < 2.0.0"
   required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 3.0"
+    }
     sym = {
       source  = "symopsio/sym"
       version = ">= 1.13.2, < 2.0.0"
